% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{check_convergence}
\alias{check_convergence}
\title{Check Convergence}
\usage{
check_convergence(results)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}
}
\value{
A cowplot containing panel of plot_free_energy(), plot_free_energy_change(), and plot_PIP()
}
\description{
\code{check_convergence} Check Convergence
}
\examples{
data(results)
check_convergence(results)

}
