% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{check_simulation_scores}
\alias{check_simulation_scores}
\title{check_simulation_scores}
\usage{
check_simulation_scores(data, results)
}
\arguments{
\item{data}{character; location of data saved by simulate_and_save e.g. simulated_true/yourname_data.Rdata"}

\item{results}{list; output of load_results()}
}
\value{
Panel of Heatmaps and scatter plots comparing the truth (data) to SDA inference (results)
}
\description{
\code{check_simulation_scores} check_simulation_scores
NB use nmf.options(grid.patch=TRUE) to stop blank pages appearing!
}
\examples{
# see vignette

}
