% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SDA.R
\name{export_data}
\alias{export_data}
\title{Export data and labels}
\usage{
export_data(matrix, path = NULL, name)
}
\arguments{
\item{matrix}{numeric matrix; the matrix you wish to save in a format readable by SDA}

\item{path}{character; the directory in which you want the data to be saved}

\item{name}{character; the name of the data, will be included in the file name}
}
\value{
Nothing, the data is saved on disk in the directory specified in the parameter 'path'.
}
\description{
\code{export_data} Export data and labels
}
\examples{


}
