% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{genome_loadings}
\alias{genome_loadings}
\title{Plot Component Loadings by Genomic Location}
\usage{
genome_loadings(component = NULL, max.items = 20, label.size = 3,
  label.repel = 1, label_both = FALSE, label_X = FALSE,
  min_loading = 0.01, gene_locations = rna_locations)
}
\arguments{
\item{component}{numeric; A named numeric vector containing the gene loadings and their names as the attribute "names"}

\item{max.items}{integer; Number of genes to label each side of 0.
I.e. if label_both is TRUE then the total number of labels will be twice max.items.}

\item{label_both}{logical; If TRUE the top N genes will be labeled for positive and negative loadings independently. 
If FALSE only genes with either positive OR negative loadings will be labeled, depending on which side has the highest loading.}

\item{gene_locations}{data.table; The output of load_gene_locations() containing the gene chromosome location coordinates}
}
\value{
A data table of statistics and their values
}
\description{
\code{genome_loadings} Plot Component Loadings by Genomic Location
}
\examples{
genome_loadings(results$loadings[[1]][48,])

}
