% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get.location}
\alias{get.location}
\title{Load genomic locations from biomart}
\usage{
get.location(gene.symbols, data_set, gene_name)
}
\arguments{
\item{gene.symbols}{character vector; Vector of gene names to look up genomic coordinates for}

\item{data_set}{character; which dataset from Ensembl should be used, e.g. "mmusculus_gene_ensembl" or "hsapiens_gene_ensembl"}

\item{gene_name}{character; which biomart value to use for matching e.g. "external_gene_name"}
}
\value{
data.table with columns: "gene_symbol", "chromosome_name" and "start_position"
}
\description{
\code{get.location} Load genome locations for a list of genes
}
