% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{highest_components}
\alias{highest_components}
\title{Which component has the highest loading for a given gene}
\usage{
highest_components(results, variable_name, top_n = 5, omic = 1)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}

\item{variable_name}{character string; name or index of variable e.g. gene}

\item{top_n}{integer; the number of components to label}

\item{omic}{integer or character; index of the omic type to calculate contributions for, or name of omic type if you assigned a name}
}
\value{
A ggplot2 object
}
\description{
\code{which_component} plot component loadings for a given gene
}
\examples{
data(results)
highest_components(results, "Xrn1")

}
