% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{highest_genes}
\alias{highest_genes}
\title{Which gene has the highest loading for a given component}
\usage{
highest_genes(results, component, omic = 1, max.items = 20,
  label.size = 3, label.repel = 1)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}

\item{component}{character string; name or index of component}

\item{omic}{integer or character; index of the omic type to calculate contributions for, or name of omic type if you assigned a name}

\item{max.items}{integer; the number of components to label}

\item{label.size}{numeric; size of point labels}

\item{label.repel}{numeric; force of label repulsion}
}
\value{
A ggplot2 object
}
\description{
\code{which_gene} plot gene loadings for a given component
}
\examples{
data(results)
highest_genes(results, 8)

}
