% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{load_chromosome_lengths}
\alias{load_chromosome_lengths}
\title{Load chromosome lengths}
\usage{
load_chromosome_lengths(organism = "mmusculus_gene_ensembl")
}
\arguments{
\item{organism}{character; which dataset from Ensembl should be used, e.g. "mmusculus_gene_ensembl" or "hsapiens_gene_ensembl"}
}
\value{
A data table with columns: chromosome, length, length_padded, genomic_offset, center
}
\description{
Load chromosome lengths
}
