% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{load_gene_locations}
\alias{load_gene_locations}
\title{Load gene coordinates from Ensembl biomart or Cached data}
\usage{
load_gene_locations(genes = NULL, cache = TRUE, path = "",
  organism = "mmusculus_gene_ensembl", name = "mouse")
}
\arguments{
\item{genes}{character vector; Vector of gene names to look up genomic coordinates for}

\item{cache}{logical; If TRUE a cache of the downloaded locations will be saved for quick loading next time}

\item{path}{character; string of path where the cached location file should be saved}

\item{organism}{character; which dataset from Ensembl should be used, e.g. "mmusculus_gene_ensembl" or "hsapiens_gene_ensembl"}

\item{name}{character; descriptive name of dataset, this will be appended to the standard cache file name when saving / loading the cache}
}
\value{
A data table of genes chromosome and coordinate. An object named chromosome.lengths will also be created in the global environment.
}
\description{
\code{load_gene_locations} Load gene coordinates from Ensembl biomart or Cached data
}
\examples{
data(results)
gene_locations <- load_gene_locations(colnames(results$loadings[[1]]))

}
