% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PIP_threshold_distribution}
\alias{PIP_threshold_distribution}
\title{Plot faction PIP < 0.5 distribution}
\usage{
PIP_threshold_distribution(results, omic = 1)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}

\item{omic}{integer or character; index of the omic type to calculate contributions for, or name of omic type if you assigned a name}
}
\value{
A ggplot2 object
}
\description{
\code{PIP_threshold_distribution} Plot faction PIP < 0.5 distribution
}
\examples{


}
