% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_free_energy}
\alias{plot_free_energy}
\title{Plot free energy over time}
\usage{
plot_free_energy(results)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}
}
\value{
A ggplot2 object
}
\description{
\code{plot_free_energy} Plot free energy over time
}
\examples{


}
