% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot loadings matrices}
\usage{
plot_loadings(data, data_type)
}
\arguments{
\item{data}{character; location of data saved by simulate_and_save e.g. simulated_true/yourname_data.Rdata"}

\item{data_type}{numeric; index of component to plot loadings for}
}
\value{
Panel of Heatmaps of loadings
}
\description{
\code{plot_loadings} plot_loadings
NB use nmf.options(grid.patch=TRUE) to stop blank pages appearing!
}
\examples{


}
