% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_PIP}
\alias{plot_PIP}
\title{Plot PIP fraction <0.5 by iteration}
\usage{
plot_PIP(results, burn_in = 50)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}

\item{burn_in}{integer; number of iterations to skip plotting at the start}
}
\value{
A ggplot2 object
}
\description{
\code{plot_PIP} Plot PIP fraction <0.5 by iteration
}
\examples{
data(results)
plot_PIP(results)

}
