% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_PIP_change}
\alias{plot_PIP_change}
\title{Plot change in fraction of PIPs <0.5 over time}
\usage{
plot_PIP_change(results)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}
}
\value{
A ggplot2 object
}
\description{
\code{plot_PIP_change} Plot change in fraction of PIPs <0.5 over time
}
\examples{
data(results)
plot_PIP_change(results)

}
