% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_scores}
\alias{plot_scores}
\title{Plot Individual Scores}
\usage{
plot_scores(results, component)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}

\item{component}{integer; The component index to plot scores for}
}
\value{
A ggplot2 object
}
\description{
\code{plot_scores} Plot Individual Scores
}
\examples{


}
