% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_scree}
\alias{plot_scree}
\title{Scree Plot}
\usage{
plot_scree(results, omic = 1)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}

\item{omic}{numeric or character; index of the omic type to calculate contributions for, or name of omic type if you assigned a name}
}
\value{
A ggplot2 object
}
\description{
\code{plot_scree} Plot scree
}
\examples{


}
