% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_500_gene_names}
\alias{random_500_gene_names}
\title{Random gene names}
\format{A charachter vector containing names of genes:}
\usage{
random_500_gene_names
}
\description{
A random vector of 500 gene names
}
\keyword{datasets}
