% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{results}
\alias{results}
\title{Results from running SDA on simulated 2D dataset}
\format{A list containing parameter estimates and run information including:
\describe{
  \item{scores}{individual scores matrix}
  \item{loadings}{list of gene loadings matrices for each omic/data type}
}}
\usage{
results
}
\description{
A dataset containing the results of SDA inference on a simulated 2D dataset
}
\keyword{datasets}
