% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{reversed log scale}
\usage{
reverselog_trans(base = 10, breaks = 5)
}
\arguments{
\item{base;}{what base log to use, e.g. 10, exp(1) (Default)}
}
\description{
reversed log scale
}
