% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SDA.R
\name{run_SDA}
\alias{run_SDA}
\title{run_SDA}
\usage{
run_SDA(sda_location = "sda", out = NULL, data = NULL, N = NULL,
  num_comps = 10, max_iter = 5000, save_freq = 5001, free_freq = 50,
  ignore_missing = FALSE, remove_zero_comps = TRUE, num_blocks = 1,
  num_openmp_threads = 1)
}
\arguments{
\item{sda_location}{string; location of the sda executable e.g. "../SDA/build/sda"
the default value 'sda' assumes the location of sda is in your $PATH variable}

\item{All}{the other parameters are identical to those for the command line bash arguments. Please see the SDA manual for the descriptions.
NB Not all parameters are implemented here.}
}
\value{
Nothing, the results are saved on disk in the directory specified in the parameter 'out'.
}
\description{
\code{run_SDA} Run SDA
}
\examples{


}
