% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scores_distribution}
\alias{scores_distribution}
\title{Plot overall distribution of individual scores}
\usage{
scores_distribution(results)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}
}
\value{
A ggplot2 object
}
\description{
\code{scores_distribution} Plot overall distribution of individual scores
}
\examples{


}
