% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_2D_data}
\alias{simulate_2D_data}
\title{Simulate 2D Data}
\usage{
simulate_2D_data(n_individuals = 100, n_variables = 500,
  n_components = 10, sparsity = 0.9)
}
\arguments{
\item{n_individuals}{integer; the number of individuals to simualte}

\item{n_variables}{integer vector; vector of integers specifying the number of variables (e.g. genes) in each data type}

\item{n_components}{integer; number of true components to simulate from}

\item{sparsity}{numeric; value between 0 and 1 for the sparsity}
}
\value{
list of Y, A, X and noise matrices
}
\description{
\code{simulate_2D_data} Simulate 2D Data
}
\examples{


}
