% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{loading_distribution}
\alias{loading_distribution}
\title{Plot overall distribution of gene loadings}
\usage{
loading_distribution(results, omic = 1)
}
\arguments{
\item{results}{list; object containing result  (output of read_output)}

\item{omic}{numeric or character; index of the omic type to plot, or name of omic type if you assigned a name}
}
\value{
A ggplot2 object
}
\description{
\code{loading_distribution} Plot overall distribution of gene loadings
}
\examples{


}
